+++ src/app/(pages)/signUp/page.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import SignUp from '@/views/signUp/ui/signUp';
+
+const page = () => {
+  return (
+    <div>
+      <SignUp />
+    </div>
+  );
+};
+
+export default page;
