--- src/app/(pages)/signup/page.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import SignUp from '@/views/signup/ui/signUp';
-
-const page = () => {
-  return <SignUp />;
-};
-
-export default page;
