+++ src/app/api/auth/sms/route.ts
@@ -13,16 +13,12 @@ export async function GET(request: Request) {
     });
     return NextResponse.json(response.data);
   } catch (error) {
-    const axiosError = error as AxiosError<{ message: string }>;
-
-    const status = axiosError.response?.status || 500;
-    const message =
-      axiosError.response?.data?.message || '코드 인증에 실패 했습닌다.';
-
-    return NextResponse.json({ error: message, status }, { status });
+    return NextResponse.json(
+      { error: 'Code verification failed' },
+      { status: 400 },
+    );
   }
 }
-
 export async function POST(request: Request) {
   const body = await request.json();
 
@@ -33,9 +29,8 @@ export async function POST(request: Request) {
     const axiosError = error as AxiosError<{ message: string }>;
 
     const status = axiosError.response?.status || 500;
-    const message =
-      axiosError.response?.data?.message || '문자 매시지 전송에 실패했습니다.';
+    const message = axiosError.response?.data?.message || 'sms failed';
 
-    return NextResponse.json({ error: message, status }, { status });
+    return NextResponse.json({ error: message }, { status });
   }
 }
