+++ src/entities/name-tag/ui/NameTagHeader/index.tsx
@@ -14,9 +14,10 @@ const NameTagHeader = ({ params }: NameTagHeaderProps) => {
     <div className="flex justify-between">
       <p className="text-h3 text-black">참가자 명찰 인원</p>
       <div className="w-[213px]">
-        <Button onClick={() => openModal('현장 신청 QR 코드')}>
-          현장 QR 조회하기
-        </Button>
+        <Button
+          onClick={() => openModal('현장 신청 QR 코드')}
+          text="현장 QR 조회하기"
+        />
       </div>
       {isModalOpen && (
         <ModalLayout>
