+++ src/entities/signUp/model/useSignup.ts
@@ -1,8 +1,15 @@
 import { useMutation } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import { toast } from 'react-toastify';
-import { SignUpData } from '@/shared/types/signup/type';
 import { postSignup } from '../api/postSignup';
+
+interface SignUpData {
+  name: string;
+  nickname: string;
+  email: string;
+  password: string;
+  phoneNumber: string;
+}
 
 export const useSignup = () => {
   const router = useRouter();
@@ -13,8 +20,8 @@ export const useSignup = () => {
       toast.success('회원가입이 완료되었습니다.');
       router.push('/signIn');
     },
-    onError: (error: Error) => {
-      toast.error(error.message);
+    onError: () => {
+      toast.error('회원가입에 실패했습니다.');
     },
   });
 };
