+++ src/entities/signUp/model/useTimer.ts
