--- src/entities/signup/index.tsx
@@ -1,5 +1,0 @@
-export { default as EmailInput } from './ui/EmailInput';
-export { default as NameInput } from './ui/NameInput';
-export { default as NicknameInput } from './ui/NicknameInput';
-export { default as PasswordInput } from './ui/PasswordInput';
-export { default as PhoneVerification } from './ui/PhoneVerification';
