--- src/entities/signup/ui/NicknameInput/index.tsx
@@ -1,26 +1,0 @@
-import React from 'react';
-import { UseFormRegister, FieldErrors } from 'react-hook-form';
-import { SignUpData } from '@/shared/types/signup/type';
-import { Input } from '@/shared/ui';
-
-interface NicknameInputProps {
-  register: UseFormRegister<SignUpData>;
-  errors: FieldErrors<SignUpData>;
-}
-
-const NicknameInput = ({ register, errors }: NicknameInputProps) => (
-  <div className="space-y-8">
-    <p className="text-h4 text-black">아이디</p>
-    <Input
-      {...register('nickname', { required: '아이디를 입력해주세요.' })}
-      type="text"
-      placeholder="아이디를 입력해주세요."
-      error={!!errors.nickname}
-    />
-    {errors.nickname && (
-      <p className="text-caption2 text-error">{errors.nickname.message}</p>
-    )}
-  </div>
-);
-
-export default NicknameInput;
