--- src/shared/types/signup/type.ts
@@ -1,9 +1,0 @@
-export interface SignUpData {
-  name: string;
-  nickname: string;
-  email: string;
-  password: string;
-  checkPassword: string;
-  phoneNumber: string;
-  code: string;
-}
