+++ src/shared/ui/Button/index.stories.ts
@@ -8,32 +8,24 @@ const meta: Meta<typeof Button> = {
 };
 
 export default meta;
-
 type Story = StoryObj<typeof Button>;
 
 export const Default: Story = {
   args: {
-    children: '확인',
-    variant: 'default',
-    disabled: false,
-    width: '100%',
+    text: '확인',
   },
 };
 
 export const White: Story = {
   args: {
-    children: '확인',
-    variant: 'white',
-    disabled: false,
-    width: '100%',
+    style: 'white',
+    text: '확인',
   },
 };
 
 export const Main100: Story = {
   args: {
-    children: '확인',
-    variant: 'main100',
-    disabled: false,
-    width: '100%',
+    style: 'main100',
+    text: '확인',
   },
 };
