+++ src/views/form/ui/FormEditor/index.tsx
@@ -81,9 +81,11 @@ const FormEditor = ({
             })
           }
         />
-        <Button type="submit" disabled={isLoading || isSuccess}>
-          {isLoading ? '제출 중...' : isSuccess ? '완료됨' : '다음'}
-        </Button>
+        <Button
+          type="submit"
+          text={isLoading ? '제출 중...' : isSuccess ? '완료됨' : '다음'}
+          disabled={isLoading || isSuccess}
+        />
       </form>
     </div>
   );
