+++ src/widgets/SMS/ui/SMSContainer/index.tsx
@@ -62,9 +62,10 @@ export default function Write() {
           />
         </div>
         <div className="w-full mobile:px-5">
-          <Button disabled={isSubmitting || sendSMSMutation.isPending}>
-            보내기
-          </Button>
+          <Button
+            disabled={isSubmitting || sendSMSMutation.isPending}
+            text="보내기"
+          />
         </div>
       </div>
     </form>
