+++ src/widgets/expo-detail/index.tsx
@@ -1,2 +1,1 @@
 export { default as ExpoDetailLayout } from './ui/ExpoDetailLayout';
-export { default as ExpoActionPanel } from './ui/ExpoActionPanel';
