+++ src/widgets/expo-detail/ui/ExpoActionPanel/index.tsx
@@ -20,9 +20,10 @@ const ExpoActionPanel = ({ params }: ExpoActionPanelProps) => {
   const getButtons = () => {
     if (role === 'user') {
       return (
-        <Button onClick={() => openModal('해당 박람회에 지원하시겠습니까?')}>
-          신청하기
-        </Button>
+        <Button
+          onClick={() => openModal('해당 박람회에 지원하시겠습니까?')}
+          text="신청하기"
+        />
       );
     }
     if (role === 'manage') {
@@ -33,40 +34,35 @@ const ExpoActionPanel = ({ params }: ExpoActionPanelProps) => {
               onClick={() => {
                 router.push(`/name-tag/${params}`);
               }}
-            >
-              QR 조회하기
-            </Button>
+              text="QR 조회하기"
+            />
             <Button
               onClick={() => openModal('누구에게 문자를 전송하시겠습니까?')}
-            >
-              문자 보내기
-            </Button>
+              text="문자 보내기"
+            />
           </div>
           <Button
             onClick={() => {
               router.push(`/program/${params}?navigation=standard`);
             }}
-            variant="main100"
-          >
-            프로그램
-          </Button>
+            text="프로그램"
+            style="main100"
+          />
 
           <Button
             onClick={() => {
               router.push(`/expo-manage/${params}`);
             }}
-            variant="main100"
-          >
-            조회하기
-          </Button>
+            text="조회하기"
+            style="main100"
+          />
           <Button
             onClick={() => {
               router.push(`/exhibition/edit/${params}`);
             }}
-            variant="white"
-          >
-            수정하기
-          </Button>
+            text="수정하기"
+            style="white"
+          />
         </div>
       );
     }
