+++ src/widgets/signUp/index.tsx
@@ -1,0 +1,1 @@
+export { default as SignUpContainer } from './ui/SignUpContainer';
